rename: move package: test -> main (#9)
