+++ src/main/java/com/shyoon/wms/location/domain/LocationRepository.java
@@ -1,11 +1,11 @@
-package com.shyoon.wms.location.feature;
+package com.shyoon.wms.location.domain;
 
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-class LocationRepository {
+public class LocationRepository {
 
     private final Map<Long, Location> locations = new HashMap<>();
     private Long sequence = 1L;
